forge fmt (#168)
* use forge fmt

* format

+++ packages/world/src/systems/DropSystem.sol
@@ -3,21 +3,21 @@ pragma solidity >=0.8.24;
 
 import { System } from "@latticexyz/world/src/System.sol";
 
+import { ActionType } from "../codegen/common.sol";
 import { ObjectType } from "../codegen/tables/ObjectType.sol";
 import { ObjectTypeMetadata } from "../codegen/tables/ObjectTypeMetadata.sol";
 import { Position } from "../codegen/tables/Position.sol";
 import { ReversePosition } from "../codegen/tables/ReversePosition.sol";
-import { ActionType } from "../codegen/common.sol";
 
+import { EntityId } from "../EntityId.sol";
 import { ObjectTypeId } from "../ObjectTypeId.sol";
 import { ObjectTypes } from "../ObjectTypes.sol";
 import { getUniqueEntity } from "../Utils.sol";
-import { transferInventoryNonEntity, transferInventoryEntity } from "../utils/InventoryUtils.sol";
-import { notify, DropNotifData } from "../utils/NotifUtils.sol";
-import { TerrainLib } from "./libraries/TerrainLib.sol";
-import { EntityId } from "../EntityId.sol";
 import { Vec3 } from "../Vec3.sol";
 import { getOrCreateEntityAt } from "../utils/EntityUtils.sol";
+import { transferInventoryEntity, transferInventoryNonEntity } from "../utils/InventoryUtils.sol";
+import { DropNotifData, notify } from "../utils/NotifUtils.sol";
+import { TerrainLib } from "./libraries/TerrainLib.sol";
 
 // TODO: combine the tool and non-tool drop functions
 contract DropSystem is System {
@@ -36,19 +36,14 @@ contract DropSystem is System {
     transferInventoryNonEntity(callerEntityId, entityId, ObjectType._get(entityId), dropObjectTypeId, numToDrop);
 
     notify(
-      callerEntityId,
-      DropNotifData({ dropCoord: coord, dropObjectTypeId: dropObjectTypeId, dropAmount: numToDrop })
+      callerEntityId, DropNotifData({ dropCoord: coord, dropObjectTypeId: dropObjectTypeId, dropAmount: numToDrop })
     );
   }
 
   function dropTool(EntityId callerEntityId, EntityId toolEntityId, Vec3 coord) public {
     EntityId entityId = dropCommon(callerEntityId, coord);
-    ObjectTypeId toolObjectTypeId = transferInventoryEntity(
-      callerEntityId,
-      entityId,
-      ObjectType._get(entityId),
-      toolEntityId
-    );
+    ObjectTypeId toolObjectTypeId =
+      transferInventoryEntity(callerEntityId, entityId, ObjectType._get(entityId), toolEntityId);
 
     notify(callerEntityId, DropNotifData({ dropCoord: coord, dropObjectTypeId: toolObjectTypeId, dropAmount: 1 }));
   }
@@ -59,13 +54,9 @@ contract DropSystem is System {
     EntityId entityId = dropCommon(callerEntityId, coord);
 
     ObjectTypeId toolObjectTypeId;
-    for (uint i = 0; i < toolEntityIds.length; i++) {
-      ObjectTypeId currentToolObjectTypeId = transferInventoryEntity(
-        callerEntityId,
-        entityId,
-        ObjectTypes.Air,
-        toolEntityIds[i]
-      );
+    for (uint256 i = 0; i < toolEntityIds.length; i++) {
+      ObjectTypeId currentToolObjectTypeId =
+        transferInventoryEntity(callerEntityId, entityId, ObjectTypes.Air, toolEntityIds[i]);
       if (i > 0) {
         require(toolObjectTypeId == currentToolObjectTypeId, "All tools must be of the same type");
       } else {
